requirements.txt
nats-py
sqlalchemy
psycopg2-binary
python-dotenv
